+++ src/utils.hpp
@@ -15,15 +15,18 @@
 #include <string>
 #include <iostream>
 #include <fstream>
+#include <vector>
+
+#include "gif.h"
 
 /* Takes a state of cells and outputs it to a PPM file. */
-void output_frame(int width, int height, std::string filename, int* cells, 
+void output_frame(int width, int height, std::string filename, uint8_t* cells, 
                   int iter_num);
 
 /* Takes a state of cells and outputs it to standard output. */
-void print_cells(int width, int height, int* cells);
+void print_cells(int width, int height, uint8_t* cells);
 
 /* Loads a state of cells from a file. */
-int* load_cells(int width, int height, std::string filename);
+uint8_t* load_cells(int width, int height, std::string filename);
 
 #endif
